dataset/requirements.txt
decorator==5.2.1
ffmpeg-python==0.2.0
future==1.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pip==25.2
proglog==0.1.12
python-dotenv==1.2.1
setuptools==78.1.1
tqdm==4.67.1
wheel==0.45.1